Rename mi.core.time to mi.core.time_tools to prevent collision with time module
