+++ mi/instrument/nortek/driver.py
@@ -47,7 +47,7 @@ from mi.core.exceptions import InstrumentProtocolException
 from mi.core.exceptions import InstrumentParameterException
 from mi.core.exceptions import SampleException
 
-from mi.core.time import get_timestamp_delayed
+from mi.core.time_tools import get_timestamp_delayed
 from mi.core.common import BaseEnum
 
 # newline.
