+++ mi/instrument/rbr/xr_420_thermistor_24/ooicore/driver.py
@@ -19,7 +19,7 @@ import ntplib
 import struct
 
 from mi.core.common import BaseEnum
-from mi.core.time import get_timestamp_delayed
+from mi.core.time_tools import get_timestamp_delayed
 from mi.core.instrument.instrument_driver import DriverParameter
 
 from mi.core.instrument.instrument_protocol import CommandResponseInstrumentProtocol
