+++ mi/instrument/satlantic/suna_deep/ooicore/driver.py
@@ -39,7 +39,7 @@ from mi.core.exceptions import InstrumentProtocolException
 from mi.core.exceptions import InstrumentParameterException
 from mi.core.exceptions import InstrumentException
 
-from mi.core.time import get_timestamp_delayed
+from mi.core.time_tools import get_timestamp_delayed
 
 # newline.
 NEWLINE = '\r\n'
