+++ mi/instrument/seabird/sbe54tps/driver.py
@@ -18,7 +18,7 @@ log = get_logger()
 
 from mi.core.util import dict_equal
 from mi.core.common import BaseEnum
-from mi.core.time import get_timestamp_delayed
+from mi.core.time_tools import get_timestamp_delayed
 from mi.core.instrument.instrument_fsm import InstrumentFSM
 from mi.core.instrument.instrument_driver import DriverEvent
 from mi.core.instrument.instrument_driver import DriverAsyncEvent
