+++ mi/instrument/star_asimet/bulkmet/metbk_a/driver.py
@@ -22,7 +22,7 @@ from mi.core.log import get_logger ; log = get_logger()
 from mi.core.common import BaseEnum
 from mi.core.exceptions import SampleException, \
                                InstrumentProtocolException
-from mi.core.time import get_timestamp_delayed
+from mi.core.time_tools import get_timestamp_delayed
                                
 from mi.core.instrument.instrument_protocol import CommandResponseInstrumentProtocol
 from mi.core.instrument.instrument_fsm import ThreadSafeFSM
